+++ src/helpers/Action.ts
@@ -1,0 +1,3 @@
+interface Action<T> {
+    (item: T): void;
+}